refactor: edge drawers are now also in the correct (Cairo or Matplotlib) submodule
