+++ src/igraph/drawing/baseclasses.py
@@ -3,6 +3,9 @@ Abstract base classes for the drawing routines.
 """
 
 from abc import ABCMeta, abstractmethod
+from math import atan2, pi
+
+from .text import TextAlignment
 
 #####################################################################
 
@@ -70,6 +73,116 @@ class AbstractXMLRPCDrawer(AbstractDrawer):
         url_parts = list(url_parts)
         url_parts[1] = hostname
         return urlunparse(url_parts)
+
+
+#####################################################################
+
+
+class AbstractEdgeDrawer(metaclass=ABCMeta):
+    """Abstract edge drawer object from which all concrete edge drawer
+    implementations are derived.
+    """
+
+    @staticmethod
+    def _curvature_to_float(value):
+        """Converts values given to the 'curved' edge style argument
+        in plotting calls to floating point values."""
+        if value is None or value is False:
+            return 0.0
+        if value is True:
+            return 0.5
+        return float(value)
+
+    @abstractmethod
+    def draw_directed_edge(self, edge, src_vertex, dest_vertex):
+        """Draws a directed edge.
+
+        @param edge: the edge to be drawn. Visual properties of the edge
+          are defined by the attributes of this object.
+        @param src_vertex: the source vertex. Visual properties are defined
+          by the attributes of this object.
+        @param dest_vertex: the source vertex. Visual properties are defined
+          by the attributes of this object.
+        """
+        raise NotImplementedError
+
+    @abstractmethod
+    def draw_undirected_edge(self, edge, src_vertex, dest_vertex):
+        """Draws an undirected edge.
+
+        @param edge: the edge to be drawn. Visual properties of the edge
+          are defined by the attributes of this object.
+        @param src_vertex: the source vertex. Visual properties are defined
+          by the attributes of this object.
+        @param dest_vertex: the source vertex. Visual properties are defined
+          by the attributes of this object.
+        """
+        raise NotImplementedError
+
+    def get_label_position(self, edge, src_vertex, dest_vertex):
+        """Returns the position where the label of an edge should be drawn. The
+        default implementation returns the midpoint of the edge and an alignment
+        that tries to avoid overlapping the label with the edge.
+
+        @param edge: the edge to be drawn. Visual properties of the edge
+          are defined by the attributes of this object.
+        @param src_vertex: the source vertex. Visual properties are given
+          again as attributes.
+        @param dest_vertex: the target vertex. Visual properties are given
+          again as attributes.
+        @return: a tuple containing two more tuples: the desired position of the
+          label and the desired alignment of the label, where the position is
+          given as C{(x, y)} and the alignment is given as C{(horizontal, vertical)}.
+          Members of the alignment tuple are taken from constants in the
+          L{TextAlignment} class.
+        """
+        # TODO: curved edges don't play terribly well with this function,
+        # we could try to get the mid point of the actual curved arrow
+        # (Bezier curve) and use that.
+
+        # Determine the angle of the line
+        dx = dest_vertex.position[0] - src_vertex.position[0]
+        dy = dest_vertex.position[1] - src_vertex.position[1]
+        if dx != 0 or dy != 0:
+            # Note that we use -dy because the Y axis points downwards
+            angle = atan2(-dy, dx) % (2 * pi)
+        else:
+            angle = None
+
+        # Determine the midpoint
+        pos = (
+            (src_vertex.position[0] + dest_vertex.position[0]) / 2.0,
+            (src_vertex.position[1] + dest_vertex.position[1]) / 2,
+        )
+
+        # Determine the alignment based on the angle
+        pi4 = pi / 4
+        if angle is None:
+            halign, valign = TextAlignment.CENTER, TextAlignment.CENTER
+        else:
+            index = int((angle / pi4) % 8)
+            halign = [
+                TextAlignment.RIGHT,
+                TextAlignment.RIGHT,
+                TextAlignment.RIGHT,
+                TextAlignment.RIGHT,
+                TextAlignment.LEFT,
+                TextAlignment.LEFT,
+                TextAlignment.LEFT,
+                TextAlignment.LEFT,
+            ][index]
+            valign = [
+                TextAlignment.BOTTOM,
+                TextAlignment.CENTER,
+                TextAlignment.CENTER,
+                TextAlignment.TOP,
+                TextAlignment.TOP,
+                TextAlignment.CENTER,
+                TextAlignment.CENTER,
+                TextAlignment.BOTTOM,
+            ][index]
+
+        return pos, (halign, valign)
 
 
 #####################################################################
