+++ src/igraph/drawing/cairo/__init__.py
@@ -1,3 +1,3 @@
 from .plot import CairoPlot
 
-__all__ = ("CairoPlot", )
+__all__ = ("CairoPlot",)
